# design/api_docs.yml
openapi: 3.0.3
info:
  title: Music Player
  description: |-
  version: 1.0.01
  contact:
    email: bsharathofficial@gmail.com
servers:
  - url: http://localhost:5000
...
paths:
  /songs/details:
    get:
      tags:
        - song
      summary: Finds song details by song id
      description: find songs
      parameters:
        - name: song_id
          in: query
          description: song id of the song
          required: false
          explode: true
          schema:
            type: string
            default: None

      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Song'          
        '400':
          description: Invalid status value
  /songs/rating:
    post:
      tags:
        - song
      summary: Add rating for a song
      description: User can rate 1-5 for a song
      requestBody:
        description: Rating object
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: integer
                  example: 1
                song_id:
                  type: integer
                  example: 1
                rating:
                  type: integer
                  example: 3
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: string
                example: Song rated succesfully
    get:
      tags:
        - song
      summary: Returns avg rating of a song
      description: Returns avg rating of a song
      parameters:
        - name: song_id
          in: query
          description: song id of the song
          required: false
          explode: true
          schema:
            type: string
            default: None
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                      type: object
                      properties:
                        rating_info:
                          type: object
                          properties:
                            rating_avg:
                              type: number
                              example: 4.5
                            rating_count:
                              type: integer
                              example: 5
                        song_id:
                          type: integer
                          example: 1
      security:
        - api_key: []
  /songs/search:
    get:
      tags:
        - song
      summary: Search bar
      description: Search by song, artist, album, genre
      parameters:
        - name: type
          in: query
          description: song/artist/album/genre
          required: true
          schema:
            type: string
        - name: value
          in: query
          description: value for the entity
          required: true
          schema:
            type: string

      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  songs:
                    type: array
                    items:
                      $ref: '#components/schemas/Song'
...
components:
  schemas:
...
    Song:
      type: object
      properties:
        song_id:
          type: integer
          format: int64
          example: 1
        song_name:
          type: string
          example: hosana
        album_name:
          type: string
          example: vtv
        artist_name:
          type: string
          example: ar rahman
        genre_name:
          type: string
          example: melody
      xml:
        name: pet
...